# docs/api-specification.yaml
openapi: 3.0.0
info:
  title: ToDo API
  description: FlaskとSQLiteを使用したToDoアプリのRESTful API（JWT認証機能付き）
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
    description: 開発サーバー

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWTトークンを使用した認証。'Authorization'ヘッダーに'Bearer {token}'形式で指定してください。

  schemas:
    # ユーザー関連のスキーマ
    User:
      type: object
      required:
        - username
        - email
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          description: ユーザーの一意の識別子
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: ユーザー名
        password:
          type: string
          minLength: 8
          maxLength: 200
          description: パスワード(DBに保存される際はハッシュ化)
        email:
          type: string
          format: email
          description: ユーザーのメールアドレス
        created_at:
          type: string
          format: date-time
          readOnly: true
          description: ユーザーが作成されたタイムスタンプ
      example:
        id: 1
        username: "testuser"
        email: "test@example.com"
        created_at: "2025-04-04T12:00:00Z"
    
    UserRegister:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: ユーザー名
        email:
          type: string
          format: email
          description: ユーザーのメールアドレス
        password:
          type: string
          format: password
          minLength: 8
          description: パスワード（8文字以上）
      example:
        username: "newuser"
        email: "newuser@example.com"
        password: "securepassword123"
    
    UserLogin:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: ユーザー名またはメールアドレス
        password:
          type: string
          format: password
          description: パスワード
      example:
        username: "testuser"
        password: "securepassword123"
    
    TokenResponse:
      type: object
      required:
        - access_token
        - refresh_token
      properties:
        access_token:
          type: string
          description: JWTアクセストークン
        refresh_token:
          type: string
          description: JWTリフレッシュトークン
        token_type:
          type: string
          default: "bearer"
          description: トークンタイプ
        expires_in:
          type: integer
          description: アクセストークンの有効期間（秒）
      example:
        access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type: "bearer"
        expires_in: 3600
    
    # ToDo関連のスキーマ（既存）
    Todo:
      type: object
      required:
        - title
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          description: ToDoアイテムの一意の識別子
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: ToDoアイテムのタイトル
        description:
          type: string
          nullable: true
          maxLength: 500
          description: ToDoアイテムの詳細説明
        due_date:
          type: string
          format: date
          nullable: true
          description: ToDoアイテムの期限日 (YYYY-MM-DD)
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
          description: ToDoアイテムの優先度
        status:
          type: string
          enum: [not_started, in_progress, completed]
          default: not_started
          description: ToDoアイテムの現在の状態
        created_at:
          type: string
          format: date-time
          readOnly: true
          description: ToDoアイテムが作成されたタイムスタンプ
        updated_at:
          type: string
          format: date-time
          readOnly: true
          description: ToDoアイテムが最後に更新されたタイムスタンプ
        user_id:
          type: integer
          format: int64
          readOnly: true
          description: ToDoアイテムの所有者ID
      example:
        id: 1
        title: "OpenAPI仕様書の作成"
        description: "ToDo APIのための包括的なOpenAPI仕様を作成する"
        due_date: "2025-04-30"
        priority: "high"
        status: "in_progress"
        created_at: "2025-04-04T12:00:00Z"
        updated_at: "2025-04-04T12:00:00Z"
        user_id: 1
    
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: HTTPステータスコード
        message:
          type: string
          description: エラーメッセージ
        details:
          type: object
          description: 追加のエラー詳細
          nullable: true
      example:
        code: 404
        message: "ToDoが見つかりません"
    
    ValidationError:
      type: object
      required:
        - code
        - message
        - errors
      properties:
        code:
          type: integer
          format: int32
          description: HTTPステータスコード
        message:
          type: string
          description: エラーメッセージ
        errors:
          type: object
          description: フィールド別のバリデーションエラー
      example:
        code: 400
        message: "バリデーションエラー"
        errors:
          title: "タイトルは必須です"
          priority: "優先度は次のいずれかである必要があります: low, medium, high"
    
    DatabaseError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: HTTPステータスコード
        message:
          type: string
          description: エラーメッセージ
        error_type:
          type: string
          description: データベースエラーの種類
      example:
        code: 500
        message: "データベースエラーが発生しました"
        error_type: "接続に失敗しました"
    
    AuthError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: HTTPステータスコード
        message:
          type: string
          description: エラーメッセージ
      example:
        code: 401
        message: "認証に失敗しました"
  
  responses:
    NotFound:
      description: 指定されたリソースが見つかりませんでした
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            TodoNotFound:
              value:
                code: 404
                message: "ID 123のToDoは見つかりません"
    
    BadRequest:
      description: リクエストに無効なパラメータまたはペイロードが含まれています
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          examples:
            MissingTitle:
              value:
                code: 400
                message: "バリデーションエラー"
                errors:
                  title: "タイトルは必須です"
            InvalidPriority:
              value:
                code: 400
                message: "バリデーションエラー"
                errors:
                  priority: "優先度は次のいずれかである必要があります: low, medium, high"
            InvalidDateFormat:
              value:
                code: 400
                message: "バリデーションエラー"
                errors:
                  due_date: "無効な日付形式です。期待される形式: YYYY-MM-DD"
            TitleTooLong:
              value:
                code: 400
                message: "バリデーションエラー"
                errors:
                  title: "タイトルは100文字を超えることができません"
    
    Conflict:
      description: 競合のためリクエストを完了できませんでした
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            DuplicateTodo:
              value:
                code: 409
                message: "同じタイトルのToDoアイテムが既に存在します"
            DuplicateUsername:
              value:
                code: 409
                message: "このユーザー名は既に使用されています"
            DuplicateEmail:
              value:
                code: 409
                message: "このメールアドレスは既に使用されています"
    
    Unauthorized:
      description: 認証情報が不足しているか無効です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthError'
          examples:
            MissingToken:
              value:
                code: 401
                message: "認証トークンがありません"
            InvalidToken:
              value:
                code: 401
                message: "無効な認証トークンです"
            ExpiredToken:
              value:
                code: 401
                message: "認証トークンの有効期限が切れています"
    
    Forbidden:
      description: 要求されたリソースへのアクセスが禁止されています
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            NotOwner:
              value:
                code: 403
                message: "このリソースにアクセスする権限がありません"
    
    InternalError:
      description: 内部サーバーエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            DatabaseConnectionFailed:
              value:
                code: 500
                message: "データベースへの接続に失敗しました"
            UnexpectedError:
              value:
                code: 500
                message: "予期しないエラーが発生しました"

paths:
  # 認証関連のエンドポイント
  /auth/register:
    post:
      summary: 新規ユーザー登録
      description: 新しいユーザーアカウントを作成します
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: ユーザーが正常に登録されました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  
  /auth/login:
    post:
      summary: ユーザーログイン
      description: ユーザー認証を行い、JWTトークンを発行します
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: ログインに成功しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: 認証に失敗しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
              example:
                code: 401
                message: "ユーザー名またはパスワードが正しくありません"
        '500':
          $ref: '#/components/responses/InternalError'
  
  /auth/refresh:
    post:
      summary: トークンの更新
      description: リフレッシュトークンを使用して新しいアクセストークンを取得します
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  description: リフレッシュトークン
      responses:
        '200':
          description: トークンの更新に成功しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  
  /auth/logout:
    post:
      summary: ユーザーログアウト
      description: 現在のセッションからログアウトし、トークンを無効化します
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ログアウトに成功しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "ログアウトに成功しました"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  
  /auth/profile:
    get:
      summary: ユーザープロファイル取得
      description: 現在認証されているユーザーのプロファイル情報を取得します
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ユーザープロファイル
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  
  # ToDo関連のエンドポイント（認証要件を追加）
  /todos:
    get:
      summary: すべてのToDoアイテムを取得
      description: 現在のユーザーのすべてのToDoアイテムのリストを返します
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: ステータスによるToDoのフィルタリング
          required: false
          schema:
            type: string
            enum: [not_started, in_progress, completed]
        - name: priority
          in: query
          description: 優先度によるToDoのフィルタリング
          required: false
          schema:
            type: string
            enum: [low, medium, high]
        - name: due_before
          in: query
          description: 指定された日付より前に期限のあるToDoをフィルタリング (YYYY-MM-DD)
          required: false
          schema:
            type: string
            format: date
        - name: due_after
          in: query
          description: 指定された日付より後に期限のあるToDoをフィルタリング (YYYY-MM-DD)
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: ToDoアイテムのリスト
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '400':
          description: 無効なクエリパラメータ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
              examples:
                InvalidStatus:
                  value:
                    code: 400
                    message: "バリデーションエラー"
                    errors:
                      status: "ステータスは次のいずれかである必要があります: not_started, in_progress, completed"
                InvalidDateFormat:
                  value:
                    code: 400
                    message: "バリデーションエラー"
                    errors:
                      due_before: "無効な日付形式です。期待される形式: YYYY-MM-DD"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    
    post:
      summary: 新しいToDoアイテムを作成
      description: 提供されたデータで現在のユーザーの新しいToDoアイテムを作成します
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 100
                  description: ToDoアイテムのタイトル
                description:
                  type: string
                  nullable: true
                  maxLength: 500
                  description: ToDoアイテムの詳細説明
                due_date:
                  type: string
                  format: date
                  nullable: true
                  description: ToDoアイテムの期限日 (YYYY-MM-DD)
                priority:
                  type: string
                  enum: [low, medium, high]
                  description: ToDoアイテムの優先度レベル
                status:
                  type: string
                  enum: [not_started, in_progress, completed]
                  description: ToDoアイテムの現在の状態
      responses:
        '201':
          description: ToDoが正常に作成されました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  
  /todos/{id}:
    parameters:
      - name: id
        in: path
        description: ToDoアイテムのID
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    
    get:
      summary: 特定のToDoアイテムを取得
      description: IDによって特定のToDoアイテムを返します
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ToDoアイテム
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: 無効なID形式
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
              example:
                code: 400
                message: "バリデーションエラー"
                errors:
                  id: "IDは正の整数である必要があります"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    
    put:
      summary: ToDoアイテムを更新
      description: 提供されたデータでToDoアイテムを更新します
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 100
                  description: ToDoアイテムのタイトル
                description:
                  type: string
                  nullable: true
                  maxLength: 500
                  description: ToDoアイテムの詳細説明
                due_date:
                  type: string
                  format: date
                  nullable: true
                  description: ToDoアイテムの期限日 (YYYY-MM-DD)
                priority:
                  type: string
                  enum: [low, medium, high]
                  description: ToDoアイテムの優先度レベル
                status:
                  type: string
                  enum: [not_started, in_progress, completed]
                  description: ToDoアイテムの現在の状態
      responses:
        '200':
          description: ToDoが正常に更新されました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
    
    delete:
      summary: ToDoアイテムを削除
      description: IDによってToDoアイテムを削除します
      security:
        - bearerAuth: []
      responses:
        '204':
          description: ToDoが正常に削除されました
        '400':
          description: 無効なID形式
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
              example:
                code: 400
                message: "バリデーションエラー"
                errors:
                  id: "IDは正の整数である必要があります"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'